pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Konstantin Kiselev <konstantin.kiselev.1998@mail.ru>"]
readme = "README.md"
repository = 'https://github.com/Kostyanuch-c/python-project-50.git'
classifiers = [
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Intended Audience :: Developers',
    'Programming Language :: Python',
]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
